pyproject.toml
[project]
name = "driftnote"
version = "0.1.0"
description = "Email-driven personal journaling app"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "Maciej Makowski" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "apscheduler>=3.10",
    "jinja2>=3.1",
    "pillow>=11.0",
    "pillow-heif>=0.20",
    "aioimaplib>=2.0",
    "aiosmtplib>=3.0",
    "structlog>=24.4",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "pyjwt[crypto]>=2.9",
    "typer>=0.13",
    "markdown-it-py>=3.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
driftnote = "driftnote.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/driftnote"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "hypothesis>=6.115",
    "freezegun>=1.5",
    "testcontainers>=4.8",
    "ruff>=0.7",
    "mypy>=1.13",
    "pre-commit>=4.0",
    "types-requests",
    "types-pyyaml",
]

[tool.ruff]
line-length = 100
target-version = "py314"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",
    "F",
    "I",
    "N",
    "B",
    "UP",
    "RUF",
    "PTH",
    "SIM",
    "S",
]
ignore = [
    "E501",
    "B008",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
plugins = ["pydantic.mypy"]
mypy_path = "src"
files = ["src", "tests"]

[[tool.mypy.overrides]]
module = ["aioimaplib.*", "apscheduler.*", "pillow_heif.*", "testcontainers.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "live: requires real Gmail credentials; not run in CI",
    "slow: deselect with '-m \"not slow\"'",
]
addopts = "-ra --strict-markers"
